Add a basic component generator
This integrates testing with ava and
adds a simple component generator module.

It uses one of the new template strings to
build up the string for the component. Though,
a lot of the variable preparation could, and
should, be broken into utils so that they
can be shared among generators.